Added max-size
Added security against large files. the maximum amount of bytes has been set to 1500000, or +- 1,5 mb.
Also conferted every space in the filename with an underscore.

+++ pages/docent/gebruikersprofiel.php
@@ -3,19 +3,24 @@
 		if(!empty($_FILES['profpic']['name'])){
 			$f_type = $_FILES['profpic']['type'];
 			if ($f_type== "image/gif" OR $f_type== "image/png" OR $f_type== "image/jpeg" OR $f_type== "image/JPEG" OR $f_type== "image/PNG" OR $f_type== "image/GIF"){
-				if($_SESSION['pf_Pic'] != "gebruikersBestanden/profilePictures/blank-profile-picture.png"){
-					unlink($_SESSION['pf_Pic']);
+				if($_FILES['profpic']['size'] <= 1500000){
+					if($_SESSION['pf_Pic'] != "gebruikersBestanden/profilePictures/blank-profile-picture.png"){
+						unlink($_SESSION['pf_Pic']);
+					}
+					$UploadedFileName = str_replace(" ","_", $_FILES['profpic']['name']);
+					$upload_directory = "gebruikersBestanden/profilePictures/"; //This is the folder which you created just now
+					$time = time();
+					$TargetPath=$time.$UploadedFileName;
+				
+					if(move_uploaded_file($_FILES['profpic']['tmp_name'], $upload_directory.$TargetPath)){ 
+						$db->updateProfielFoto($_SESSION['user_id'], $upload_directory.$TargetPath);
+						$_SESSION['pf_Pic'] = $upload_directory.$TargetPath;
+						echo "<script>window.location.href='gebruikersprofiel';</script>";
+						exit;
+					}
 				}
-				$UploadedFileName = $_FILES['profpic']['name'];
-				$upload_directory = "gebruikersBestanden/profilePictures/"; //This is the folder which you created just now
-				$time = time();
-				$TargetPath=$time.$UploadedFileName;
-			
-				if(move_uploaded_file($_FILES['profpic']['tmp_name'], $upload_directory.$TargetPath)){ 
-					$db->updateProfielFoto($_SESSION['user_id'], $upload_directory.$TargetPath);
-					$_SESSION['pf_Pic'] = $upload_directory.$TargetPath;
-					echo "<script>window.location.href='gebruikersprofiel';</script>";
-					exit;
+				else{
+					$message = $lang['BIG_FILE'];
 				}
 			}
 			else{
